+++ react/.storybook/main.js
@@ -1,19 +1,14 @@
+/** @type { import('@storybook/react-vite').StorybookConfig } */
 module.exports = {
   stories: [
-    '../stories/**/*.stories.mdx',
     '../stories/**/*.stories.@(js|jsx|ts|tsx)',
   ],
   addons: [
     '@storybook/addon-links',
     '@storybook/addon-essentials',
-    '@storybook/addon-knobs/register',
   ],
-
-  webpackFinal: async (config) => {
-    config.plugins = config.plugins.filter(
-      (plugin) => plugin.constructor.name !== 'ProgressPlugin'
-    )
-
-    return config
+  framework: {
+    name: '@storybook/react-vite',
+    options: {},
   },
 }